+++ lib/showcaseview.dart
@@ -23,6 +23,7 @@
 library showcaseview;
 
 export 'src/models/action_button_icon.dart';
+export 'src/models/progress_indicator_config.dart';
 export 'src/models/tooltip_action_button.dart';
 export 'src/models/tooltip_action_config.dart';
 export 'src/showcase/showcase.dart';
@@ -30,4 +31,6 @@ export 'src/showcase/showcase_view.dart';
 export 'src/showcase_widget.dart';
 export 'src/utils/enum.dart' hide ShowcaseProgressType, TooltipLayoutSlot;
 export 'src/widget/floating_action_widget.dart';
+export 'src/widget/showcase_floating_action_widget.dart';
+export 'src/widget/showcase_progress_indicator.dart';
 export 'src/widget/tooltip_action_button_widget.dart';